@@ -1,34 +1,39 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AssemblyName>thZero.NetCore.Library.Asp.Security.NWebSec</AssemblyName>
-    <PackageId>thZero.NetCore.Library.Asp.Security.NWebSec</PackageId>
-    <Description></Description>
-    <Authors>thZero.com</Authors>
-    <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
-    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
-    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
-    <GenerateAssemblyCopyrightAttribute>false</GenerateAssemblyCopyrightAttribute>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
-    <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/thzero/thZero.NetCore.Library.Asp.Security.NWebSec</PackageProjectUrl>
-    <RepositoryType>GitHub</RepositoryType>
-    <RepositoryUrl>https://github.com/thzero/thZero.NetCore.Library.Asp.Security.NWebSec</RepositoryUrl>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net5.0</TargetFramework>
+		<AssemblyName>thZero.NetCore.Library.Asp.Security.NWebSec</AssemblyName>
+		<PackageId>thZero.NetCore.Library.Asp.Security.NWebSec</PackageId>
+		<Description></Description>
+		<Authors>thZero.com</Authors>
+		<GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
+		<GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
+		<GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
+		<GenerateAssemblyCopyrightAttribute>false</GenerateAssemblyCopyrightAttribute>
+		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+		<PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
+		<PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
+		<PackageProjectUrl>https://github.com/thzero/thZero.NetCore.Library.Asp.Security.NWebSec</PackageProjectUrl>
+		<RepositoryType>GitHub</RepositoryType>
+		<RepositoryUrl>https://github.com/thzero/thZero.NetCore.Library.Asp.Security.NWebSec</RepositoryUrl>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <Compile Include="..\thZero.NetCore.Library\CommonAssemblyInfo.cs" Link="CommonAssemblyInfo.cs" />
-  </ItemGroup>
+	<ItemGroup>
+		<Compile Include="..\thZero.NetCore.Library\CommonAssemblyInfo.cs" Link="CommonAssemblyInfo.cs" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="NWebsec.AspNetCore.Middleware" Version="2.0.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<FrameworkReference Include="Microsoft.AspNetCore.App" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\thZero.NetCore.Library.Asp\thZero.NetCore.Library.Asp.csproj" />
-    <ProjectReference Include="..\thZero.NetCore.Library\thZero.NetCore.Library.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+	  <PackageReference Include="NWebsec.AspNetCore.Middleware" Version="3.0.0" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\thZero.NetCore.Library.Asp.Instrumentation\thZero.NetCore.Library.Asp.Instrumentation.csproj" />
+	  <ProjectReference Include="..\thZero.NetCore.Library.Asp\thZero.NetCore.Library.Asp.csproj" />
+	  <ProjectReference Include="..\thZero.NetCore.Library\thZero.NetCore.Library.csproj" />
+	</ItemGroup>
 
 </Project>
